Create README.md
I made a Circuit for pressure control in a hyperbaric chamber using several electronic devices to perform the specific functions. Thus, I needed a sensor, a repeater, a domain converter, a comparator and a relay + red LED assembly.
To find out all the necessary information, please follow the documentation that provides all the necessary indications.